--- webui/src/App.css
@@ -1,12 +1,0 @@
-.app-container {
-  font-family: 'Arial', sans-serif;
-  color: #333333;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-}
-
-.page-content {
-  flex-grow: 1;
-  padding: 24px;
-}
