+++ webui/src/styles/Dashboard.css
@@ -1,0 +1,13 @@
+.app-container {
+    font-family: 'Arial', sans-serif;
+    color: #333333;
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+  }
+  
+  .page-content {
+    flex-grow: 1;
+    padding: 24px;
+  }
+
